add organizations resource

+++ apps/api/src/organizations/organizations.service.ts
@@ -1,13 +1,14 @@
-import { Injectable } from '@nestjs/common';
-import { PrismaService } from '../prisma.service';
+import { Injectable, NotFoundException } from '@nestjs/common';
+
 import type {
   CreateOrganizationDto,
   UpdateOrganizationDto,
   Organization,
 } from '@repo/api';
+import { PrismaService } from '../prisma.service';
 
 @Injectable()
-export class OrganizationService {
+export class OrganizationsService {
   constructor(private prisma: PrismaService) {}
 
   async create(
@@ -22,16 +23,30 @@ export class OrganizationService {
     return this.prisma.organization.findMany();
   }
 
-  async findOne(id: string): Promise<Organization | null> {
-    return this.prisma.organization.findUnique({
+  async findOne(id: string): Promise<Organization> {
+    const organization = await this.prisma.organization.findUnique({
       where: { id },
     });
+
+    if (!organization) {
+      throw new NotFoundException(`Organization with ID ${id} not found`);
+    }
+
+    return organization;
   }
 
   async update(
     id: string,
     updateOrganizationDto: UpdateOrganizationDto,
   ): Promise<Organization> {
+    const organization = await this.prisma.organization.findUnique({
+      where: { id },
+    });
+
+    if (!organization) {
+      throw new NotFoundException(`Organization with ID ${id} not found`);
+    }
+
     return this.prisma.organization.update({
       where: { id },
       data: updateOrganizationDto,
@@ -39,6 +54,14 @@ export class OrganizationService {
   }
 
   async remove(id: string): Promise<Organization> {
+    const organization = await this.prisma.organization.findUnique({
+      where: { id },
+    });
+
+    if (!organization) {
+      throw new NotFoundException(`Organization with ID ${id} not found`);
+    }
+
     return this.prisma.organization.delete({
       where: { id },
     });
